More refactoring, fixed some issues when base_url has trailing slash

+++ ugs/inbox.py
@@ -5,6 +5,7 @@ import requests
 from flask import Blueprint, request, jsonify, make_response
 
 from ugs.activitypub.signature import sign_and_send
+from ugs.models.activity import Activity
 from ugs.models.db import db
 from ugs.models.actor import Actor
 from ugs.models.follower import Follower
@@ -15,10 +16,12 @@ bp = Blueprint('inbox', __name__, url_prefix='/user/<username>/inbox')
 
 
 def handle_follow(req, username):
-    actor_obj = Actor.query.filter_by(steam_name=username).first()
+    actor_obj = Actor.query.filter_by(ugs_id=username).first()
     if actor_obj is None:
         print("Actor not found")
-        return "Actor not found", 404
+        response = jsonify({'error': 'Actor not found'})
+        response.status_code = 404
+        return response
 
     ap_object = req['object']
     activity_type = req['type']
@@ -31,7 +34,9 @@ def handle_follow(req, username):
         print("Fetching foreign actor")
         actor_request = requests.get(external_actor, headers={'Accept': 'application/activity+json'})
         if actor_request.status_code != 200:
-            return "Failed to fetch foreign actor", 400
+            response = jsonify({'error': 'Failed to fetch foreign actor'})
+            response.status_code = 400
+            return response
         foreign_actor_obj = actor_request.json()
 
         # Store the foreign actor in the database
@@ -81,7 +86,6 @@ def handle_follow(req, username):
     print("BASE URL: ", base_url)
 
     activity_id = f"{base_url}/activities/{accept_guid}"
-    accept_url = f"{base_url}/user/{actor_obj['steam_name']}/inbox/{accept_guid}"
     accept = {
         '@context': 'https://www.w3.org/ns/activitystreams',
         'type': 'Accept',
@@ -102,12 +106,18 @@ def handle_follow(req, username):
     )
 
     # Store the activity in the database
-    new_activity = ForeignActivity(
-        activity_id=str(accept_guid),
+    #     db.execute(
+    #         'INSERT INTO activity (guid, actor_guid, activity_type, object_guid, activity_json) VALUES (?, ?, ?, ?, ?)',
+    #         (str(accept_guid), actor_obj['steam_name'], 'Accept', foreign_activity_id, str(accept))
+    #     )
+    #     db.commit()
+    new_activity = Activity(
+        guid=str(accept_guid),
         actor_guid=actor_obj.ugs_id,
         activity_type='Accept',
         object_guid=foreign_activity_id,
-        activity_json=str(accept)
+        activity_json=str(accept),
+        screenshot_id=None
     )
     db.session.add(new_activity)
     db.session.commit()
@@ -128,10 +138,17 @@ def handle_follow(req, username):
 def inbox(username):
     print(f"Received request for {username}'s inbox")
     # Handles AP requests to the inbox
-    actor_obj = Actor.query.filter_by(steam_name=username).first()
+    actor_obj = Actor.query.filter_by(ugs_id=username).first()
 
     if actor_obj is None:
-        return "Actor not found", 404
+        print("Actor not found")
+        print("Username:", username)
+        # Print first row of the actor table
+        actor_obj = Actor.query.first()
+        print("First row of actor table:", actor_obj)
+        response = jsonify({'error': 'Actor not found'})
+        response.status_code = 404
+        return response
 
     ap_object = request.json.get('object')
     activity_type = request.json['type']
@@ -139,11 +156,15 @@ def inbox(username):
     foreign_activity_id = request.json['id']
 
     if activity_type is None:
-        return "Missing activity type", 400
+        print("Missing activity type")
+        response = jsonify({'error': 'Missing activity type'})
+        response.status_code = 400
+        return response
 
     response = None
     match activity_type:
         case 'Follow':
+            print("handle follow activity")
             response = handle_follow(request.json, username)
         case 'Undo':
             print("Undo activity")
@@ -151,7 +172,12 @@ def inbox(username):
             print("AP Object:", ap_object)
             # Undo activity
             if ap_object['type'] == 'Follow':
-                follower = Follower.query.filter_by(follower_id=external_actor, following_id=actor_obj['ugs_id']).first()
+                follower = Follower.query.filter_by(follower_id=external_actor, following_id=actor_obj.ugs_id).first()
+                if follower is None:
+                    print("Follower not found")
+                    response = jsonify({'error': 'Follower not found'})
+                    response.status_code = 404
+                    return response
                 db.session.delete(follower)
                 db.session.commit()
             else:
